Moved bodyInscripcionClase.
